# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QAA'A API",
        "version": "1.0.0",
        "description": "API for managing hall bookings and reservations"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token"
            }
        }
    },
    "paths": {
...
        "/api/bookings/{id}/confirm": {
            "post": {
                "tags": ["Bookings"],
                "summary": "Confirm booking payment",
                "operationId": "confirmPayment",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed"
                    }
                }
            }
        },
...
    }
}